refactor: move TxnOptions to dedicated file for better code organization
